Add <releases> and OARS data
I am actually not sure this OARS data is right. After you complete a
level, the game shows you a random article from the UN's Universal
Declaration of Human Rights. (They're distributed as images, not text,
and in the tarball we're using here they're further packed into some
archive format I don't recognise.)

If, like me, you need to refresh your memory, the Declaration is here:
https://www.un.org/en/universal-declaration-human-rights/

Article 4 is:

> No one shall be held in slavery or servitude; slavery and the slave
> trade shall be prohibited in all their forms.

This is arguably a reference to historical slavery. Should this gain a
violence-slavery: mild tag? Probably yes, but this would cause
appstream-glib to assign this game a minimum age of 13. It's been a long
time since I was 13, but I think this game is more suitable for a rather
younger audience…

Fixes #1